+++ libs/translator/writer/write_light.cpp
@@ -40,27 +40,68 @@ PXR_NAMESPACE_USING_DIRECTIVE
 
 namespace {
 
-bool writeArnoldLightShader(const AtNode* node, UsdPrim& prim, 
-    UsdArnoldPrimWriter &primWriter, UsdArnoldWriter &writer, const TfToken& lightShaderAttr)
+UsdPrim _GetNodeGraph(UsdPrim& prim, UsdArnoldWriter &writer, const TfToken& lightShaderAttr)
 {
-    AtNode *linkedColor = AiNodeGetLink(node, str::color);
-    if (!linkedColor)
-        return false;
     std::string nodeGraphName = prim.GetPath().GetString() + "/light_shader";
     SdfPath nodeGraphPath(nodeGraphName);        
     UsdStageRefPtr stage = writer.GetUsdStage();
     UsdPrim nodeGraphPrim = stage->DefinePrim(nodeGraphPath, str::t_ArnoldNodeGraph);
-    
+
     UsdAttribute arnoldShaderAttr = 
         prim.CreateAttribute(lightShaderAttr, 
         SdfValueTypeNames->String, false);
     arnoldShaderAttr.Set(nodeGraphPrim.GetPath().GetString());
 
+    return nodeGraphPrim;
+}
+
+bool _WriteArnoldLightShader(const AtNode* node, UsdPrim& prim, 
+    UsdArnoldPrimWriter &primWriter, UsdArnoldWriter &writer, const TfToken& lightShaderAttr)
+{
+    AtNode *linkedColor = AiNodeGetLink(node, str::color);
+    
+    if (!linkedColor)
+        return false;
+    
+    UsdPrim nodeGraphPrim = _GetNodeGraph(prim, writer, lightShaderAttr);
+    
     primWriter.WriteAttribute(node, "color", nodeGraphPrim, 
         nodeGraphPrim.CreateAttribute(str::t_outputs_color, 
             SdfValueTypeNames->Token, false), writer);
     return true;
 }
+
+bool _WriteArnoldLightFilters(const AtNode* node, UsdPrim& prim, 
+    UsdArnoldPrimWriter &primWriter, UsdArnoldWriter &writer, const TfToken& lightShaderAttr)
+{
+    const AtArray *lightFilters = AiNodeGetArray(node, str::filters);    
+    unsigned int numFilters = lightFilters ? AiArrayGetNumElements(lightFilters) : 0;
+
+    if (numFilters == 0)
+        return false;
+
+    UsdPrim nodeGraphPrim = _GetNodeGraph(prim, writer, lightShaderAttr);
+    for (unsigned int i = 0; i < numFilters; ++i) {
+        AtNode* filter = (AtNode*)AiArrayGetPtr(lightFilters, i);
+        if (filter == nullptr)
+            continue;
+
+        TfToken filterIndexAttrName(TfStringPrintf("outputs:filters:i%d", i+1));
+        UsdAttribute filterIndexAttr = nodeGraphPrim.CreateAttribute(filterIndexAttrName, 
+            SdfValueTypeNames->Token, false);
+        
+        writer.WritePrimitive(filter);
+        std::string filterName = UsdArnoldPrimWriter::GetArnoldNodeName(filter, writer);
+        UsdPrim filterPrim = writer.GetUsdStage()->GetPrimAtPath(SdfPath(filterName));
+        UsdAttribute filterOutputAttr = filterPrim.CreateAttribute(str::t_outputs_out, SdfValueTypeNames->Token, false);
+        filterName += ".outputs:out";
+        
+        SdfPath filterOutputPath(filterName);
+        filterIndexAttr.AddConnection(filterOutputPath);
+    }
+    return true;
+}
+
 
 void writeLightCommon(const AtNode *node, UsdPrim &prim, UsdArnoldPrimWriter &primWriter, UsdArnoldWriter &writer)
 {
@@ -94,7 +135,9 @@ void UsdArnoldWriteDistantLight::Write(const AtNode *node, UsdArnoldWriter &writ
 
     UsdAttribute normalizeAttr = prim.CreateAttribute(TfToken("primvars:arnold:normalize"), SdfValueTypeNames->Bool, false);
     WriteAttribute(node, "normalize", prim, normalizeAttr, writer);
-
+    
+    _WriteArnoldLightFilters(node, prim, *this, writer, str::t_primvars_arnold_shaders);
+    _exportedAttrs.insert("filters");
     _WriteArnoldParameters(node, writer, prim, "primvars:arnold");
 }
 
@@ -109,8 +152,8 @@ void UsdArnoldWriteDomeLight::Write(const AtNode *node, UsdArnoldWriter &writer)
 
     writeLightCommon(node, prim, *this, writer);
     _WriteMatrix(light, node, writer);
-
-    if (writeArnoldLightShader(node, prim, *this, writer, str::t_primvars_arnold_shaders)) {
+    
+    if (_WriteArnoldLightShader(node, prim, *this, writer, str::t_primvars_arnold_shaders)) {
         _exportedAttrs.insert("color");
     
         AtNode *linkedTexture = AiNodeGetLink(node, "color");
@@ -125,6 +168,9 @@ void UsdArnoldWriteDomeLight::Write(const AtNode *node, UsdArnoldWriter &writer)
             writer.SetAttribute(light.GetColorAttr(), GfVec3f(1.f, 1.f, 1.f));
         }
     }
+    _WriteArnoldLightFilters(node, prim, *this, writer, str::t_primvars_arnold_shaders);
+    _exportedAttrs.insert("filters");
+ 
     AtString textureFormat = AiNodeGetStr(node, AtString("format"));
     static AtString latlongStr("latlong");
     static AtString mirrored_ballStr("mirrored_ball");
@@ -157,6 +203,8 @@ void UsdArnoldWriteDiskLight::Write(const AtNode *node, UsdArnoldWriter &writer)
     WriteAttribute(node, "radius", prim, light.GetRadiusAttr(), writer);
     WriteAttribute(node, "normalize", prim, light.GetNormalizeAttr(), writer);
     _WriteMatrix(light, node, writer);
+    _WriteArnoldLightFilters(node, prim, *this, writer, str::t_primvars_arnold_shaders);
+    _exportedAttrs.insert("filters");
     _WriteArnoldParameters(node, writer, prim, "primvars:arnold");
 }
 
@@ -182,6 +230,8 @@ void UsdArnoldWriteSphereLight::Write(const AtNode *node, UsdArnoldWriter &write
     }
 
     _WriteMatrix(light, node, writer);
+    _WriteArnoldLightFilters(node, prim, *this, writer, str::t_primvars_arnold_shaders);
+    _exportedAttrs.insert("filters");    
     _WriteArnoldParameters(node, writer, prim, "primvars:arnold");
 }
 
@@ -199,7 +249,7 @@ void UsdArnoldWriteRectLight::Write(const AtNode *node, UsdArnoldWriter &writer)
     _WriteMatrix(light, node, writer);
     WriteAttribute(node, "normalize", prim, light.GetNormalizeAttr(), writer);
 
-    if (writeArnoldLightShader(node, prim, *this, writer, str::t_primvars_arnold_shaders)) {
+    if (_WriteArnoldLightShader(node, prim, *this, writer, str::t_primvars_arnold_shaders)) {
         _exportedAttrs.insert("color");
 
         AtNode *linkedTexture = AiNodeGetLink(node, AtString("color"));
@@ -214,6 +264,8 @@ void UsdArnoldWriteRectLight::Write(const AtNode *node, UsdArnoldWriter &writer)
             writer.SetAttribute(light.GetColorAttr(), GfVec3f(1.f, 1.f, 1.f));        
         }
     }
+    _WriteArnoldLightFilters(node, prim, *this, writer, str::t_primvars_arnold_shaders);
+    _exportedAttrs.insert("filters");
 
     float width = 1.f;
     float height = 1.f;
@@ -260,7 +312,9 @@ void UsdArnoldWriteCylinderLight::Write(const AtNode *node, UsdArnoldWriter &wri
 
     WriteAttribute(node, "normalize", prim, light.GetNormalizeAttr(), writer);
     _WriteMatrix(light, node, writer);
-    
+    _WriteArnoldLightFilters(node, prim, *this, writer, str::t_primvars_arnold_shaders);
+    _exportedAttrs.insert("filters");
+       
     _WriteArnoldParameters(node, writer, prim, "primvars:arnold");
 }
 
@@ -286,8 +340,11 @@ void UsdArnoldWriteGeometryLight::Write(const AtNode *node, UsdArnoldWriter &wri
     // We're not authoring the light matrix, so that it's consistent with the mesh
     _exportedAttrs.insert("matrix");
 
-    if (writeArnoldLightShader(node, mesh, *this, writer, str::t_primvars_arnold_light_shaders))
+    if (_WriteArnoldLightShader(node, mesh, *this, writer, str::t_primvars_arnold_light_shaders))
         _exportedAttrs.insert("color");
+
+    _WriteArnoldLightFilters(node, mesh, *this, writer, str::t_primvars_arnold_light_shaders);
+    _exportedAttrs.insert("filters");
     
     _WriteArnoldParameters(node, writer, mesh, "primvars:arnold:light");    
 }